refactor: 데이터 검증 로직을 models -> controllers로 이동
