+++ docs/README.md
@@ -67,17 +67,18 @@
 
 
 ## 리팩터링
-- [ ] html 태그 함수는 길어져도 괜찮은지? 
-
-- [x] 구매할 수 없을 때, 수량 -1 되는 것 막음 
-- [ ] productPurchaseModel 리팩터링
-- [ ] 상품 구매 탭도 다시 클릭했을 때 가져올 수 있도록 함 
-
+- [x] 검증 로직 m이냐 c냐? -> c!
 - [ ] alert를 views로 옮기기 (models-views 서로 부르면 안됨)
+- [ ] 상품 구매 탭도 다시 클릭했을 때 가져올 수 있도록 함 
+- [ ] productPurchaseModel 리팩터링
+
+- [ ] render 하는 로직 controll에 (index.js 최대한 적게, controll이 view 호출)
+- [ ] html 태그 함수는 길어져도 괜찮은지? 
 - [ ] 함수(또는 메소드)의 길이가 15라인을 넘어가지 않도록 구현한다. (eslint 속성 추가)
   - 함수(또는 메소드)가 한 가지 일만 잘 하도록 구현한다.
 - [ ] 피드백 
 
+- [x] 구매할 수 없을 때, 수량 -1 되는 것 막음 
 - [x] this.amount 명시적으로 이름 바꾸기 
 - [x] 상수화(메시지, 숫자)
 - [x] autofocus(사용성)
