+++ src/js/models/vendingMachineModel.js
@@ -1,4 +1,4 @@
-import { CHARGE, ERROR, PRICE } from '../utils/constants.js';
+import { PRICE } from '../utils/constants.js';
 
 const getRandomCoin = () => {
   return MissionUtils.Random.pickNumberInList([PRICE.FIVE_HUNDRED_WON, PRICE.ONE_HUNDRED_WON, PRICE.FIFTY_WON, PRICE.TEN_WON]);
@@ -16,23 +16,7 @@ const countCoin = (coin, coins) => {
   }
 };
 
-export const isValidCharge = chargeInput => {
-  if (chargeInput === '') {
-    alert(ERROR.CHARGE_BLANK);
-    return false;
-  }
-  if (Number(chargeInput) < CHARGE.LEAST_INPUT_PRICE) {
-    alert(ERROR.CHARGE_LEAST_TEN);
-    return false;
-  }
-  if (Number(chargeInput % PRICE.TEN_WON !== 0)) {
-    alert(ERROR.CHARGE_SHOULD_DIVIDED_INTO_TEN);
-    return false;
-  }
-  return true;
-};
-
-export const chargeMoney = (chargeInput, coins) => {
+const chargeMoney = (chargeInput, coins) => {
   coins.amount += chargeInput;
   while (chargeInput) {
     const coin = getRandomCoin();
@@ -44,3 +28,5 @@ export const chargeMoney = (chargeInput, coins) => {
   }
   return coins;
 };
+
+export default chargeMoney;
